Add EuclideanAlgorithm.py and NthRoot.py
EuclideanAlgorithm.py defines the function EuclideanAlgorithm(A,B) which
returns the GCD of A and B and two coeffiecients satisfying a Bézout
identity for A and B. It uses the Euclidean algorithm.

NthRoot.py defines a function NthRoot(Base,N) that computes the N-th
root of Base using the bisection method. It is useful when Base is so
big that other methods such as Base ** (1/N) fails.
Result is the largest integer X such that X ** N <= Base.